swift/BrowserStackTests/ditto_sync_test.py
#!/usr/bin/env python3
"""
BrowserStack iOS Appium test for Ditto Swift Tasks app
Tests real device functionality and Ditto sync operations
"""

import os
import sys
import time
import json
from appium import webdriver
from appium.options.ios import XCUITestOptions
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import TimeoutException, NoSuchElementException

class DittoSwiftTasksTest:
    def __init__(self):
        self.driver = None
        self.test_results = {"passed": 0, "failed": 0, "tests": []}
        
    def setup_driver(self, app_url, device_config):
        """Setup BrowserStack iOS driver"""
        options = XCUITestOptions()
        options.platform_name = 'iOS'
        options.device_name = device_config['device']
        options.platform_version = device_config['os_version']
        options.app = app_url
        options.project_name = "Ditto Swift Tasks"
        options.build_name = f"Swift iOS Build {os.getenv('GITHUB_RUN_ID', 'local')}"
        options.session_name = f"Ditto Sync Test - {device_config['device']}"
        options.browser_stack_debug = True
        options.browser_stack_console = "verbose"
        options.browser_stack_network_logs = True
        options.automation_name = "XCUITest"
        options.new_command_timeout = 300
        options.wait_for_idle_timeout = 10
        
        # BrowserStack credentials
        username = os.getenv('BROWSERSTACK_USERNAME')
        access_key = os.getenv('BROWSERSTACK_ACCESS_KEY')
        
        if not username or not access_key:
            raise ValueError("BrowserStack credentials not found")
            
        remote_url = f"https://{username}:{access_key}@hub-cloud.browserstack.com/wd/hub"
        
        print(f"🚀 Starting test on {device_config['device']} (iOS {device_config['os_version']})")
        
        self.driver = webdriver.Remote(
            command_executor=remote_url,
            options=options
        )
        
        return self.driver
    
    def test_app_launch(self):
        """Test 1: App launches successfully"""
        print("🧪 Test 1: App Launch")
        try:
            # Wait for app to launch and show main screen
            WebDriverWait(self.driver, 30).until(
                lambda driver: len(driver.find_elements(By.XPATH, "//*")) > 5
            )
            
            # Take screenshot for verification
            self.driver.save_screenshot("app_launch.png")
            
            print("✅ App launched successfully")
            self.add_test_result("App Launch", "PASSED", "App launched and UI elements visible")
            return True
            
        except TimeoutException:
            print("❌ App failed to launch within 30 seconds")
            self.add_test_result("App Launch", "FAILED", "Timeout waiting for app to launch")
            return False
    
    def test_ditto_initialization(self):
        """Test 2: Ditto initializes without crashing"""
        print("🧪 Test 2: Ditto Initialization")
        try:
            # App should stay running (not crash) for at least 10 seconds
            # This indicates Ditto initialized successfully
            time.sleep(10)
            
            # Check app is still responsive
            current_source = self.driver.page_source
            if len(current_source) > 100:  # App is still showing UI
                print("✅ Ditto initialized successfully - app remains stable")
                self.add_test_result("Ditto Initialization", "PASSED", "App stable after Ditto init")
                return True
            else:
                print("❌ App appears to have crashed or lost UI")
                self.add_test_result("Ditto Initialization", "FAILED", "App lost UI after init")
                return False
                
        except Exception as e:
            print(f"❌ Ditto initialization test failed: {str(e)}")
            self.add_test_result("Ditto Initialization", "FAILED", f"Exception: {str(e)}")
            return False
    
    def test_task_operations(self):
        """Test 3: Basic task CRUD operations"""
        print("🧪 Test 3: Task Operations")
        try:
            # Look for add task button or input field
            add_elements = self.driver.find_elements(By.XPATH, "//*[contains(@name,'add') or contains(@label,'add') or @type='XCUIElementTypeTextField']")
            
            if add_elements:
                print("✅ Found task input/add elements")
                
                # Try to interact with first available element
                element = add_elements[0]
                if element.get_attribute('type') == 'XCUIElementTypeTextField':
                    # It's a text field - try to add a task
                    test_task = f"BrowserStack Test Task {int(time.time())}"
                    element.send_keys(test_task)
                    print(f"✅ Added test task: {test_task}")
                else:
                    # It's likely a button - tap it
                    element.click()
                    print("✅ Tapped add button")
                
                # Wait a moment for UI to update
                time.sleep(3)
                
                # Take screenshot of task list
                self.driver.save_screenshot("task_operations.png")
                
                self.add_test_result("Task Operations", "PASSED", "Successfully interacted with task UI")
                return True
            else:
                print("⚠️ No task input/add elements found - app may have different UI")
                self.add_test_result("Task Operations", "PARTIAL", "No task UI elements found")
                return True  # Don't fail completely - app might have different UI
                
        except Exception as e:
            print(f"❌ Task operations test failed: {str(e)}")
            self.add_test_result("Task Operations", "FAILED", f"Exception: {str(e)}")
            return False
    
    def test_app_stability(self):
        """Test 4: App remains stable during extended use"""
        print("🧪 Test 4: App Stability")
        try:
            # Keep app running for 30 seconds and check stability
            for i in range(6):
                time.sleep(5)
                
                # Check app is still responsive
                try:
                    elements = self.driver.find_elements(By.XPATH, "//*")
                    if len(elements) < 3:
                        print(f"❌ App stability issue detected at {i*5} seconds")
                        self.add_test_result("App Stability", "FAILED", f"UI elements dropped at {i*5}s")
                        return False
                    print(f"✅ App stable at {(i+1)*5} seconds ({len(elements)} UI elements)")
                except Exception:
                    print(f"❌ App became unresponsive at {i*5} seconds")
                    self.add_test_result("App Stability", "FAILED", f"Unresponsive at {i*5}s")
                    return False
            
            print("✅ App remained stable for 30 seconds")
            self.add_test_result("App Stability", "PASSED", "App stable for 30 seconds")
            return True
            
        except Exception as e:
            print(f"❌ App stability test failed: {str(e)}")
            self.add_test_result("App Stability", "FAILED", f"Exception: {str(e)}")
            return False
    
    def add_test_result(self, test_name, status, details):
        """Add test result to results tracking"""
        result = {
            "test": test_name,
            "status": status,
            "details": details,
            "timestamp": time.time()
        }
        self.test_results["tests"].append(result)
        
        if status == "PASSED":
            self.test_results["passed"] += 1
        elif status == "FAILED":
            self.test_results["failed"] += 1
    
    def run_all_tests(self, app_url, device_config):
        """Run all tests on specified device"""
        try:
            print(f"\n🎯 Starting Ditto Swift Tests on {device_config['device']}")
            print("=" * 60)
            
            # Setup driver
            self.setup_driver(app_url, device_config)
            
            # Run tests in sequence
            tests = [
                self.test_app_launch,
                self.test_ditto_initialization,
                self.test_task_operations,
                self.test_app_stability
            ]
            
            for test in tests:
                if not test():
                    print("⚠️ Test failed, but continuing with remaining tests...")
                time.sleep(2)  # Brief pause between tests
            
            print("\n📊 Test Summary:")
            print(f"✅ Passed: {self.test_results['passed']}")
            print(f"❌ Failed: {self.test_results['failed']}")
            print(f"📱 Device: {device_config['device']} (iOS {device_config['os_version']})")
            
            return self.test_results["failed"] == 0
            
        except Exception as e:
            print(f"❌ Test suite failed: {str(e)}")
            self.add_test_result("Test Suite", "FAILED", f"Suite exception: {str(e)}")
            return False
        finally:
            if self.driver:
                print("🔚 Closing BrowserStack session...")
                self.driver.quit()

def main():
    """Main test execution"""
    app_url = os.getenv('BROWSERSTACK_APP_URL')
    if not app_url:
        print("❌ BROWSERSTACK_APP_URL environment variable not set")
        sys.exit(1)
    
    # Device configurations to test
    devices = [
        {"device": "iPhone 15 Pro", "os_version": "17"},
        {"device": "iPhone 14", "os_version": "16"},
    ]
    
    overall_success = True
    results_summary = []
    
    for device_config in devices:
        print(f"\n🚀 Testing on {device_config['device']}")
        test_runner = DittoSwiftTasksTest()
        
        try:
            success = test_runner.run_all_tests(app_url, device_config)
            results_summary.append({
                "device": device_config['device'],
                "success": success,
                "results": test_runner.test_results
            })
            
            if not success:
                overall_success = False
                
        except Exception as e:
            print(f"❌ Failed to run tests on {device_config['device']}: {str(e)}")
            overall_success = False
            results_summary.append({
                "device": device_config['device'],
                "success": False,
                "error": str(e)
            })
    
    # Print overall summary
    print("\n" + "=" * 60)
    print("🏁 OVERALL TEST RESULTS")
    print("=" * 60)
    
    for result in results_summary:
        status = "✅ PASSED" if result.get('success') else "❌ FAILED"
        print(f"{result['device']}: {status}")
        
        if 'results' in result:
            print(f"  Tests: {result['results']['passed']} passed, {result['results']['failed']} failed")
    
    # Save results to file for CI
    with open('browserstack_test_results.json', 'w') as f:
        json.dump(results_summary, f, indent=2)
    
    if overall_success:
        print("\n🎉 All BrowserStack tests PASSED!")
        sys.exit(0)
    else:
        print("\n💥 Some BrowserStack tests FAILED!")
        sys.exit(1)

if __name__ == "__main__":
    main()