swift/BrowserStackTests/requirements.txt
Appium-Python-Client==4.1.0
selenium==4.15.2